requirements.txt
openai==1.12.0
python-telegram-bot==20.8
python-dotenv==1.0.1
httpx==0.27.0
